Implement typed local storage
Locals on the call stack are now associated explicitly with their
corresponding functions and are strongly typed using metadata. This
improves the performance of interacting with locals while also laying
the groundwork necessary for arbitrary data types and to align the
layout of local storage more closely with that of the heap (once it's
implemented).