Add infrastructure to support custom data types
Adds the data_type module with corresponding definitions for
TypeDefinition and the global TypeTable. Types should now be ready for
integration into the VM and execution context.

It's possible that TypeDefinition may need more metadata than it has
currently; specifically, each definition has the type of any custom
subfield erased. With the currently planned instruction set this
wouldn't be a problem for realtime operations, but it would complicate
using this information for debugging or metaprogramming. Because the
TypeDefinition is constructed during compile time, the only downside to
extending upon this format is the slight impact that it would have on
cache utilization, which is probably well worth it.