Add initial support for compound data types
Creation of compound data types using the dt_create instruction is now
implemented. A type definition table has been added to the global
context, allowing Sahara to track the structure of each data type
defined by the user. All functions that calculate value sizes or that
interact with local slots require read references to this table so that
the user's dynamically defined types can be laid out in memory alongside
Sahara's primitives.

Instructions to read and write individual fields within a created type
are forthcoming, but have been documented in the bytecode specification.