Implement read of data type fields
Individual fields can now be read from compound data types and pushed
onto the data stack. Implementation of this instruction has brought to
light the need for instruction extension (as the operation requires
information both about the local slot that must be read and the field
that must be read from within the local). Given that extension is needed
anyway, it's possible that it may make more sense to change the static
instruction size to 16 bits.

This would almost halve the instruction size of most programs. Many
instructions may not even need extension, since they would only need to
be extended if the source code required the greater number of resources.
This would mean that instructions could be marked as "extension
eligible" rather than either always being extended or not. For the
"indexed" instructions, this would simply mean that non-extended indices
are always assumed to be interpretable with only a single byte!

Need to think a bit more on this, but it feels clean and is likely to
significantly improve performance without sacrificing almost anything.